Cleanup the code sync the users between clerk and supabase

+++ src/app/api/blogs/route.ts
@@ -1,5 +1,6 @@
 import { createClerkSupabaseClientSsr } from "@/lib/supabase/server"
-import { NextRequest, NextResponse } from "next/server";
+import { auth } from "@clerk/nextjs/server";
+import { NextRequest } from "next/server";
 
 async function GET(): Promise<Response> {
   const client = await createClerkSupabaseClientSsr();
@@ -16,25 +17,28 @@ async function GET(): Promise<Response> {
     // Send the data as a JSON response
     return new Response(JSON.stringify(data), { status: 200, headers: { 'Content-Type': 'application/json' }, })
   } catch (error) {
+    console.error(error);
     return new Response(JSON.stringify({ error: "Something Went Wrong" }),
       { status: 500, headers: { 'Content-Type': 'application/json' } });
   }
 }
 
-async function POST(req: NextRequest, res: NextResponse) {
+async function POST(req: NextRequest) {
   const data = await req.json();
   const client = await createClerkSupabaseClientSsr();
   try {
     await client.from('Posts').insert([data]);
     return new Response(JSON.stringify(data), { status: 201 })
   } catch (error) {
+    console.log(error)
     return new Response(JSON.stringify({ error: "Something Went Wrong" }),
       { status: 500, headers: { 'Content-Type': 'application/json' } });
   }
 }
 
-async function PUT(req: NextRequest, res: NextResponse) {
+async function PUT(req: NextRequest) {
   const data = await req.json();
+  const {userId} = await auth();
   const client = await createClerkSupabaseClientSsr();
   try {
      const updatedData = await client
@@ -42,22 +46,26 @@ async function PUT(req: NextRequest, res: NextResponse) {
     .update({
       ...data
     })
-    .eq('id', data.id);
+    .eq('id', data.id)
+    .eq('user_id', userId);
     console.log(updatedData)
     return new Response(JSON.stringify(data), { status: 201 })
   } catch (error) {
+    console.error(error);
     return new Response(JSON.stringify({ error: "Something Went Wrong" }),
       { status: 500, headers: { 'Content-Type': 'application/json' } });
   }
 }
 
-async function DELETE(req: NextRequest, res: NextResponse) {
+async function DELETE(req: NextRequest) {
   const data = await req.json();
+  const {userId} = await auth();
   const client = await createClerkSupabaseClientSsr();
   try {
-    await client.from('Posts').delete().eq('id', data.id);;
+    await client.from('Posts').delete().eq('id', data.id).eq('user_id', userId);
     return new Response(null, { status: 201 })
   } catch (error) {
+    console.error(error);
     return new Response(JSON.stringify({ error: "Something Went Wrong" }),
       { status: 500, headers: { 'Content-Type': 'application/json' } });
   }
